TANDEM Interrupts implementation
* corev-dv end_of_test reworked to use tohost
