+++ bsp/Makefile
@@ -10,11 +10,11 @@ RISCV            ?= $(CV_SW_TOOLCHAIN)
 RISCV_EXE_PREFIX ?= $(RISCV)/bin/riscv32-unknown-elf-
 RISCV_GCC = $(RISCV_EXE_PREFIX)gcc
 RISCV_AR = $(RISCV_EXE_PREFIX)ar
-C_FILES = syscalls_kernel.c csr.c csr.h syscalls.c syscalls.h
-SRC = crt0.S handlers.S syscalls.c syscalls_kernel.c vectors.S csr.c
-OBJ = crt0.o handlers.o syscalls.o syscalls_kernel.o vectors.o csr.o
+C_FILES = syscalls_kernel.c csr.c syscalls.c syscalls.h rand.c
+SRC = crt0.S handlers.S syscalls.c syscalls_kernel.c vectors.S csr.c rand.c
+OBJ = crt0.o handlers.o syscalls.o syscalls_kernel.o vectors.o csr.o rand.o
 LIBCV-VERIF = libcv-verif.a
-CFLAGS ?= -Os -g -static -mabi=ilp32 -march=$(CV_SW_MARCH) -Wall -pedantic
+CFLAGS ?= -Os -g -static -mabi=ilp32 -march=$(RISCV_MARCH) -Wall -pedantic
 
 all: $(LIBCV-VERIF)
 
@@ -36,8 +36,8 @@ format:
 
 vars:
 	@echo "make bsp variables:"
-	@echo "   CV_SW_TOOLCHAIN  = $(CV_SW_TOOLCHAIN)"
-	@echo "   CV_SW_MARCH      = $(CV_SW_MARCH)"
+	@echo "   CV_SW_TOOLCHAIN  = $(RISCV)"
+	@echo "   CV_SW_MARCH      = $(RISCV_MARCH)"
 	@echo "   RISCV            = $(RISCV)"
 	@echo "   RISCV_EXE_PREFIX = $(RISCV_EXE_PREFIX)"
 	@echo "   RISCV_GCC        = $(RISCV_GCC)"
