--- tb/core/README.md
@@ -1,2 +1,0 @@
-# CORE testbench for CV32E20
-Coming soon...
