+++ tests/programs/custom/interrupt_test/interrupt_test.c
@@ -1,6 +1,6 @@
-//Interrupt test for CV32E20 
-//All interrupt lines are level-sensitive. It is assumed that the interrupt handler signals 
-//completion of the handling routine to the interrupt source, e.g., through a memory-mapped register, 
+//Interrupt test for CV32E20
+//All interrupt lines are level-sensitive. It is assumed that the interrupt handler signals
+//completion of the handling routine to the interrupt source, e.g., through a memory-mapped register,
 //which then deasserts the corresponding interrupt line.
 
 
@@ -9,6 +9,7 @@
 #include <stdlib.h>
 #include <time.h>
 
+#include "rand.h"
 #include "interrupt_test.h"
 
 volatile uint32_t irq_id                  = 0;
@@ -88,28 +89,17 @@ void mm_ram_assert_irq(uint32_t mask, uint32_t cycle_delay) {
     *TIMER_VAL_ADDR = 1 + cycle_delay;
 }
 
-uint32_t random_num(uint32_t upper_bound, uint32_t lower_bound) {
-    uint32_t random_num = *((volatile int *)0x15001000);
-    uint32_t num = (random_num  % (upper_bound - lower_bound + 1)) + lower_bound;
-    return num;
-}
-
-uint32_t random_num32() {
-    uint32_t num = *((volatile int *)0x15001000);
-    return num;
-}
-
 extern void __no_irq_handler();
 
 void nested_irq_handler(uint32_t id) {
     // First stack mie, mepc and mstatus
     // Must be done in critical section with MSTATUS.MIE == 0
-    volatile uint32_t mie, mepc, mip, mstatus;    
+    volatile uint32_t mie, mepc, mip, mstatus;
     asm volatile("csrr %0, mie" : "=r" (mie));
     asm volatile("csrr %0, mepc" :"=r" (mepc));
     asm volatile("csrr %0, mstatus" : "=r" (mstatus));
- 
-    // read mip 
+
+    // read mip
     asm volatile ("csrr %0,mip" : "=r" (mip));
     // Re enable interrupts and create window to enable nested irqs
     mstatus_mie_enable();
@@ -127,15 +117,15 @@ void generic_irq_handler(uint32_t id) {
     asm volatile("csrr %0, mcause": "=r" (mmcause));
     irq_id = id;
     uint32_t clear_mask = (0xFFFFFFFF - (0x1 << irq_id));
-    if (active_test ==1) { 
+    if (active_test == 1) {
       mm_ram_assert_irq(0, random_num(10,0));
     }
-    if (active_test == 2) {    
+    if (active_test == 2) {
         irq_id_q[irq_id_q_ptr++] = id;
         mm_ram_driven = (mm_ram_driven & clear_mask);
         mm_ram_assert_irq(mm_ram_driven, 8);
     //    mm_ram_assert_irq((0x1 << (32-irq_id_q_ptr)), 8);
-    } 
+    }
     if (active_test == 3 || active_test == 4) {
 // printf("generic_irq_handler: original mm_ram_driven %lx\n",  mm_ram_driven);
 // printf("generic_irq_handler: interrupt %lu %lu\n", irq_id, id);
@@ -143,16 +133,16 @@ void generic_irq_handler(uint32_t id) {
         irq_id_q[irq_id_q_ptr++] = id;
         mm_ram_driven = (mm_ram_driven & clear_mask);
         mm_ram_assert_irq(mm_ram_driven, 8);
-// printf("generic_irq_handler: mm_ram_driven %lx\n",  mm_ram_driven);
+ //printf("generic_irq_handler: mm_ram_driven %lx\n",  mm_ram_driven);
     //    mm_ram_assert_irq((0x1 << (32-irq_id_q_ptr)), 8);
-    } 
+    }
     if (active_test == 3) {
         if (nested_irq_valid) {
             nested_irq_valid = 0;
             mm_ram_assert_irq(0x1 << nested_irq, random_num(10,0));
         }
         nested_irq_handler(id);
-    }    
+    }
 }
 
 void m_software_irq_handler(void) { generic_irq_handler(SOFTWARE_IRQ_ID); }
@@ -191,7 +181,7 @@ __attribute__((interrupt ("machine"))) void u_sw_direct_irq_handler(void)  {
 	    "j __no_irq_handler\n"
 	    "j __no_irq_handler\n"
 	    "j m_software_irq_handler\n"
-	    "j __no_irq_handler\n"        
+	    "j __no_irq_handler\n"
             "j __no_irq_handler\n"
 	    "j __no_irq_handler\n"
 	    "j m_timer_irq_handler\n"
@@ -227,7 +217,7 @@ __attribute__((interrupt ("machine"))) void u_sw_direct_irq_handler(void)  {
         ".option norvc\n"
         ".align 8\n"
         "alt_direct_vector_table:\n"
-	    "j u_sw_direct_irq_handler\n"	
+	    "j u_sw_direct_irq_handler\n"
     );
 
     asm (
@@ -236,7 +226,7 @@ __attribute__((interrupt ("machine"))) void u_sw_direct_irq_handler(void)  {
         ".align 8\n"
         "alt_direct_ecall_table:\n"
         "wfi\n"
-	    "j u_sw_irq_handler\n"	
+	    "j u_sw_irq_handler\n"
     );
 
 int main(int argc, char *argv[]) {
@@ -272,7 +262,7 @@ int main(int argc, char *argv[]) {
     if (retval != EXIT_SUCCESS)
         return retval;
 
-// TODO unused tests. to be removed at clean-up 
+// TODO unused tests. to be removed at clean-up
 /*    // Repeat test1 (restore vector mode)
     retval = test7();
     if (retval != EXIT_SUCCESS)
@@ -324,13 +314,13 @@ int test1_impl(int direct_mode) {
 
                 in_direct_handler = 0;
                 mmcause = 0;
-                mm_ram_assert_irq(0x1 << i, 1);    
+                mm_ram_assert_irq(0x1 << i, 1);
 
                 if (((0x1 << i) & IRQ_MASK) && mie && gmie) {
-                    // Interrupt is valid and enabled 
+                    // Interrupt is valid and enabled
                     // wait for the irq to be served
                     while (!mmcause);
-                
+
                     if ((mmcause & (0x1 << 31)) == 0) {
                         printf("MCAUSE[31] was not set: mmcause = 0x%08lx\n", (uint32_t) mmcause);
 
@@ -339,8 +329,8 @@ int test1_impl(int direct_mode) {
                     if ((mmcause & MCAUSE_IRQ_MASK) != i) {
                         printf("MCAUSE reported wrong irq, exp = %lu, act = 0x%08lx\n", i, mmcause);
 
-                        return ERR_CODE_TEST_1;                    
-                    }                    
+                        return ERR_CODE_TEST_1;
+                    }
                 } else {
                     // Unimplemented interrupts, or is a masked irq, delay a bit, waiting for any mmcause
                     for (int j = 0; j < 20; j++) {
@@ -348,9 +338,9 @@ int test1_impl(int direct_mode) {
                             printf("MMCAUSE = 0x%08lx when unimplmented interrupt %lu first\n", mmcause, i);
                             return ERR_CODE_TEST_1;
                         }
-                    }                    
+                    }
                 }
-                
+
                 // Check MIP
                 // For unimplemented irqs, this should always be 0
                 // For masked irqs, this should always be 0
@@ -367,7 +357,7 @@ int test1_impl(int direct_mode) {
                     if (mip & (0x1 << i)) {
                         printf("MIP for unimplemented IRQ[%lu] set\n", i);
                         return ERR_CODE_TEST_1;
-                    }                    
+                    }
                 }
 
                 // Check flag at direct mode handler
@@ -383,20 +373,20 @@ int test1_impl(int direct_mode) {
                 }
 
                 // Clear vp irq
-                mm_ram_assert_irq(0, 0);   
+                mm_ram_assert_irq(0, 0);
             }
         }
     }
     mmcause = 0;
     // set NMI
-    mm_ram_assert_irq(0x1 << 0, 1);    
+    mm_ram_assert_irq(0x1 << 0, 1);
     while (!mmcause);
     if (mmcause != 0x80000020) {
         printf("MMCAUSE = 0x%08lx while expecting 0x8000001f\n", mmcause);
         return ERR_CODE_TEST_1;
     }
     // Clear NMI
-    mm_ram_assert_irq(0, 0);   
+    mm_ram_assert_irq(0, 0);
 
 
     return EXIT_SUCCESS;
@@ -409,13 +399,13 @@ int test1_impl(int direct_mode) {
 int test2() {
     printf("TEST 2 - TRIGGER ALL IRQS AT ONCE:\n");
     active_test = 2;
-    
+
     // Clear VP irq
-    mm_ram_assert_irq(0, 0); 
-    
+    mm_ram_assert_irq(0, 0);
+
     // Enable all interrupts (MIE and MSTATUS.MIE)
     uint32_t mie = (uint32_t) -1;
-    asm volatile("csrw mie, %0" : : "r" (mie));            
+    asm volatile("csrw mie, %0" : : "r" (mie));
     mstatus_mie_enable();
     irq_id_q_ptr = 0;
 
@@ -443,8 +433,8 @@ int test2() {
 
 // Test 3 will create nested interrupts
 int test3() {
-    printf("TEST 3 - NESTED INTERRUPTS:\n"); 
-    
+    printf("TEST 3 - NESTED INTERRUPTS:\n");
+
     // Test 3 is a nested interrupt test
     active_test = 3;
 
@@ -460,12 +450,13 @@ int test3() {
         uint32_t irq[2];
 
         // Pick 2 random interrupts
-        irq[0] = IRQ_ID_PRIORITY[random_num(IRQ_NUM-1, 0)];
-        irq[1] = IRQ_ID_PRIORITY[random_num(IRQ_NUM-1, 0)];
+        // Do not allow nmi
+        irq[0] = IRQ_ID_PRIORITY[random_num(IRQ_NUM-1, 1)];
+        irq[1] = IRQ_ID_PRIORITY[random_num(IRQ_NUM-1, 1)];
 
         irq_id_q_ptr = 0;
         nested_irq = irq[1];
-        nested_irq_valid = 1;    
+        nested_irq_valid = 1;
 
 #ifdef DEBUG_MSG
         printf("TEST3: Test nested irqs %lu and %lu\n", irq[0], irq[1]);
@@ -499,15 +490,15 @@ int test3() {
 // Tests that WFI works regardless of MSTATUS.MIE
 // Tests that IRQ handler is not entered after WFI unless MSTATUS.MIE is set
 int test4() {
- 
+
     printf("TEST 4 - WFI\n");
-    
+
     // Test 4 is a WFI test
     active_test = 4;
 
     // Iterate through multiple loops
     for (int irq = 1; irq < 32; irq++) {  // NMI is irq[0]
-        if (!(((0x1 << irq) & IRQ_MASK))) 
+        if (!(((0x1 << irq) & IRQ_MASK)))
             continue;
 
         for (uint32_t gmie = 0; gmie <= 1; gmie++) {
@@ -517,7 +508,7 @@ int test4() {
             mie_disable_all();
             mm_ram_assert_irq(0, 0);
 
-            // Select a wakeup interrupt and enable only it            
+            // Select a wakeup interrupt and enable only it
             mie_enable(irq);
 
             // Prep the IRQ ID Q to be empty, we need to detect if any interrupts (or none) taken
@@ -539,9 +530,9 @@ int test4() {
 
             // Random assert "enabled" irq
             mm_ram_assert_irq(rand_irq | (0x1 << irq), (random_num32() & 0x3f) + 32);
-            asm volatile("wfi");            
-            
-        
+            asm volatile("wfi");
+
+
             if (gmie) {
                 // Expected an interrupt taken
                 if (irq_id_q[0] != irq) {
@@ -559,7 +550,7 @@ int test4() {
     }
     return EXIT_SUCCESS;
 }
- 
+
 // Test 5 will repeat the basic interrupt test in test 1
 // But with a relocated vector table via mtvec CSR
 int test5() {
@@ -574,7 +565,7 @@ int test5() {
     asm volatile("csrw mtvec, %0" : : "r" ((uint32_t) alt_vector_table | (save_mtvec & 0x3)));
 
     retval = test1_impl(0);
-    asm volatile("csrw mtvec, %0" : : "r" (save_mtvec)); 
+    asm volatile("csrw mtvec, %0" : : "r" (save_mtvec));
     if (retval != EXIT_SUCCESS) {
         return ERR_CODE_TEST_5;
     }
@@ -619,7 +610,7 @@ int test6() {
 
 // Test 7 will repeat the basic interrupt test in test 1
 // But with a relocated vector table via mtvec CSR and DIRECT vector mode
-// CV32E20 does not support DIRECT mode.  
+// CV32E20 does not support DIRECT mode.
 // TODO Saved here until clean up
 /*
 int test7() {
@@ -635,7 +626,7 @@ int test7() {
     asm volatile("csrw mtvec, %0" : : "r" ((uint32_t) alt_direct_vector_table)); // Leave mode at 0
 
     retval = test1_impl(1);
-    asm volatile("csrw mtvec, %0" : : "r" (save_mtvec)); 
+    asm volatile("csrw mtvec, %0" : : "r" (save_mtvec));
     if (retval != EXIT_SUCCESS) {
         return ERR_CODE_TEST_7;
     }
@@ -657,7 +648,7 @@ int test8() {
     return EXIT_SUCCESS;
  }
  */
- 
+
 // CV32E20 - this test appears to be a coverage filler test. Will wait until we see initial coverage numbers
 /* TODO Saved here until clean up
 int test9() {
@@ -675,16 +666,16 @@ int test9() {
 
     // Iterate through multiple loops
     for (int irq = 0; irq < 32; irq++) {
-        if (!(((0x1 << irq) & IRQ_MASK))) 
+        if (!(((0x1 << irq) & IRQ_MASK)))
             continue;
 
         for (uint32_t gmie = 0; gmie <= 0; gmie++) {
             uint32_t rand_irq;
 
             // Clear MIE and all pending irqs
-            mie_disable_all();            
+            mie_disable_all();
 
-            // Select a wakeup interrupt and enable only it            
+            // Select a wakeup interrupt and enable only it
             mie_enable(irq);
 
             // Prep the IRQ ID Q to be empty, we need to detect if any interrupts (or none) taken
@@ -707,7 +698,7 @@ int test9() {
             // Random assert "enabled" irq
             mm_ram_assert_irq(rand_irq | (0x1 << irq), (random_num32() & 0x3f) + 64);
             asm volatile("ecall");
-                    
+
             if (gmie) {
                 // Expected an interrupt taken
                 if (irq_id_q[0] != irq) {
@@ -724,7 +715,7 @@ int test9() {
         }
     }
 
-    asm volatile("csrw mtvec, %0" : : "r" (save_mtvec)); 
+    asm volatile("csrw mtvec, %0" : : "r" (save_mtvec));
 
     return EXIT_SUCCESS;
 }
